requirements.txt
pandas
scikit-learn
google-generativeai
streamlit
python-dotenv
